+++ NEWS.md
@@ -5,6 +5,7 @@
  * Added support for records processing in XML format
  * Applied parallel approach for mapping from records to instances
  * Use batch post to send Instances to the mod-inventory
+ * MARC to Instance mapping was updated to the new one version
 
 ## 2019-05-17 v1.3.1
  * Filled complete date and stub HrID for Job Execution